+++ lambda/JobClean/clean_handler.py
@@ -1,6 +1,6 @@
-"""The clean lambda is used to remove all traces of the specified ipppssoots.
+"""The clean lambda is used to remove all traces of the specified datasets.
 
-If clean-all is specified,  then all files related to processing any ipppssoot
+If clean-all is specified,  then all files related to processing any dataset
 are deleted.
 """
 
@@ -9,20 +9,19 @@ from calcloud import s3
 
 
 def lambda_handler(event, context):
-
-    bucket_name, ipst = s3.parse_s3_event(event)
+    bucket_name, dataset = s3.parse_s3_event(event)
 
     comm = io.get_io_bundle(bucket_name)
 
-    if ipst == "all":
+    if dataset == "all":
         print("Cleaning all datasets;  removing all job resources (S3 files).")
 
-        comm.messages.delete_literal("clean-all")  # don't interpret all as existing ipppssoots
+        comm.messages.delete_literal("clean-all")  # don't interpret all as existing datasets
 
         cleanup_ids = comm.ids("all")
 
         comm.messages.broadcast("clean", cleanup_ids)
-    elif ipst == "ingested":  # a variation of "all" restricted to datasets with an ingest message
+    elif dataset == "ingested":  # a variation of "all" restricted to datasets with an ingest message
         print("Cleaning all ingested datasets;  removing all job resources (S3 files).")
 
         comm.messages.delete_literal("clean-ingested")  # don't interpret "ingested"
@@ -31,5 +30,5 @@ def lambda_handler(event, context):
 
         comm.messages.broadcast("clean", cleanup_ids)
     else:
-        print("Cleaning", ipst)
-        comm.clean(ipst)
+        print("Cleaning", dataset)
+        comm.clean(dataset)