+++ terraform/deploy_vars.sh
@@ -1,7 +1,7 @@
 #! /bin/bash -xu
 export CALCLOUD_VER="ami-rotate-codebuild"
-export CALDP_VER="0.2.21"
-export CAL_BASE_IMAGE="stsci/hst-pipeline:CALDP_20230208_CAL_final"
+export CALDP_VER="v0.2.22-rc3"
+export CAL_BASE_IMAGE="stsci/hst-pipeline:CALDP_onlywfpc2drizzle_CAL_rc3"
 
 export BASE_IMAGE_TAG=`cut -d ":" -f2- <<< ${CAL_BASE_IMAGE} `