+++ terraform/variables.tf
@@ -114,10 +114,10 @@ variable crds_context {
   description = "the crds context to be the default for the environment"
   type = map(string)
   default = {
-    "-sb" = "hst_0866.pmap"
-    "-dev" = "hst_0866.pmap"
-    "-test" = "hst_1015.pmap"
-    "-ops" = "hst_1015.pmap"
+    "-sb" = "hst_1074.pmap"
+    "-dev" = "hst_1074.pmap"
+    "-test" = "hst_1077.pmap"
+    "-ops" = "hst_1077.pmap"
   }
 }